forgotten debug cout's removed

+++ miheev.egor/P4/structs.hpp
@@ -13,7 +13,7 @@ namespace miheev
     int* matrix_;
 
   public:
-    Matrix (size_t rows, size_t cols, int mode);
+    Matrix (size_t rows, size_t cols, long long mode);
     void initWithIfstream(std::ifstream& in);
     void printSelf() const;
     void increaseRectBorder(size_t xStart, size_t yStart, size_t width, size_t height, int increment);
